Deserialization prototype
As of now, the deserialization routine works as expected.
It's pluggable thanks to metamodels. Nodes can specify how
they want to be serialized at the implementation's discretion.

To reduce the size of the serialized ASTs, only the most fundamental
properties of a node are saved (eg operator for a binary expression,
image of a name node).

In particular the prototype does not include text coordinates as it
truly blows up the size of the blobs. I think the only way around this
is to make PMD only about trees - not let the user query tokens
themselves

About archive size: each node takes at least 2 bytes: 1 of type ID, 1
for the children end marker.

We only have about 100 (non-deprecated) nodes in the Java AST, vs 256
possible values for the type ID. That means, that we can encode some
node "shapes" in a free type ID.

For example, there's a single node type for a Boolean literal, with a
boolean field. If we wrote out the boolean field, then every boolean
literal would be 3 bytes. However if we use JJTBOOLEANLITERAL for the
"false" literal and another free id for the "true" literal (setting the
field in the factory), we can fit the node inside 2 bytes.

Similarly (still TODO), storing the parenthesis count of every
expression with an int would punish *all expressions* with 4 bytes. In
that instance it would make sense to insert a fake node that we delete
when deserializing.

There are other things we can envision to compress the tree.
* Some nodes have always the same number of children, in which case we
could get rid of the end marker. This is free if we serialize bottom up
instead of top-down
 * in particular, some nodes are always leaves (identifiers, literals, etc)
* Identifiers have many duplicates (even if we look just within a single
source file). Outside of using a real compression algorithm, we could use
a kind of constant pool.
* We can also compress the blobs with a standard utility. gzip gives
compression rates of 500%.

For jdk12 (17300 classes):
* Space
 * Sources are 230MB
 * Serialized ASTs are 80MB
 * Compressed ASTs are 16MB
* Time
 * Parsing all sources is about 40s
 * Reconstructing all ASTs is about 3s
 * Writing the ASTs to disk takes less than 3s