Introduce ClasspathRequest for SymbolResolver calls
This represents the origin of the request. It must
be filled in properly by the caller and propagated.
Sometimes propagating to the caller is too difficult
(eg TypeTestUtil.isA) because we would have to pass
it around a lot of times. In these cases we use
unknownOrigin(), which means, all files will end up
having a dependency on this request