Add new rule NullabilityProblem
Since it isn't flow sensitive we can
only report cases where we know the
value is null always. This is already
somewhat useful but incomplete.

We should add flow sensitivity, most
likely by extending the dataflow pass
logic to visit in control flow order.
But this is very involved tbh.